requirements.txt
opencv-python
Pillow
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning
omegaconf==2.1.1
einops==0.3.0
transformers>4.19.2
open-clip-torch==2.7.0
lightning